refactor: 리팩토링
- 패키지를 변경한다.
- 메서드 이름을 수정한다.
