+++ src/main/java/lotto/support/converter/Converter.java
@@ -3,7 +3,7 @@ package lotto.support.converter;
 import java.math.BigDecimal;
 import java.util.List;
 import lotto.exception.converter.InvalidConvertException;
-import lotto.support.InputValidator;
+import lotto.support.validator.InputValidator;
 
 public class Converter {
