+++ src/main/java/lotto/support/validator/InputValidator.java
@@ -1,4 +1,4 @@
-package lotto.support;
+package lotto.support.validator;
 
 import java.util.List;
 import lotto.exception.validator.InvalidInputException;
